requirements.txt
google-generativeai>=0.8.0
python-pptx>=0.6.23
matplotlib>=3.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
mistune>=3.0.0
pandas>=2.0.0
huggingface_hub>=0.22.0